rename box_scores to player_box_scores
